+++ components/tracksManager/tacksList/TracksList.tsx
@@ -1,15 +1,20 @@
 'use client';
-import { DataTable } from './DataTable';
 import { Filters } from './Filters';
 import { Track } from '@models/zod/track.schema';
 import { TracksTableSorting } from '@models/zod/track.table.schema';
 import { useTracksListController } from '@lib/hooks/components/trackList/useTracksListController';
+import dynamic from 'next/dynamic';
+
+const DataTable = dynamic(() => import('./DataTable').then((mod) => mod.DataTable<Track, unknown, TracksTableSorting>), {
+	ssr: false,
+	loading: () => <div className="text-gray-500">Loading table...</div>,
+});
 
 export function TracksList() {
 	const { data, isLoading, error, trackTableColumns, tableSorting, tablePagination, onSortingChanged, onPageChange } =
 		useTracksListController();
 	const hasData = data && data.data.length > 0;
-
+	// console.log('Are we running on server? TracksList? : ', typeof window === 'undefined');
 	return (
 		<div className="flex flex-row gap-10">
 			<Filters />
@@ -21,7 +26,7 @@ export function TracksList() {
 			)}
 
 			{!isLoading && hasData && (
-				<DataTable<Track, unknown, TracksTableSorting>
+				<DataTable
 					columns={trackTableColumns}
 					data={data.data}
 					sorting={tableSorting}
