+++ next.config.ts
@@ -10,6 +10,18 @@ const bundleAnalyzer = withBundleAnalyzer({
 const nextConfig: NextConfig = {
 	reactStrictMode: true,
 	productionBrowserSourceMaps: true,
+	images: {
+		// Allows loading images from different domains until we add the feature to save images on the server
+		remotePatterns: [
+			{
+				protocol: 'https',
+				hostname: '*',
+				port: '',
+				search: '',
+			},
+		],
+	},
+	compress: true,
 };
 
 export default bundleAnalyzer(nextConfig);
