+++ package.json
@@ -30,8 +30,8 @@
     "axios": "^1.8.4",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
+    "date-fns": "^4.1.0",
     "lucide-react": "^0.503.0",
-    "moment": "^2.30.1",
     "neverthrow": "^8.2.0",
     "next": "15.3.1",
     "next-themes": "^0.4.6",
@@ -44,9 +44,9 @@
     "sonner": "^2.0.3",
     "tailwind-merge": "^3.2.0",
     "tw-animate-css": "^1.2.8",
+    "vitest": "^3.2.4",
     "zod": "^3.24.3",
-    "zustand": "^5.0.5",
-    "vitest": "^3.2.4"
+    "zustand": "^5.0.5"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
@@ -65,7 +65,7 @@
   "optionalDependencies": {
     "@rollup/rollup-linux-arm64-gnu": "4.44.1",
     "@rollup/rollup-linux-x64-gnu": "4.44.1",
-    "sass-embedded-linux-x64": "1.89.2",
-    "lightningcss-linux-x64-gnu": "1.30.1"
+    "lightningcss-linux-x64-gnu": "1.30.1",
+    "sass-embedded-linux-x64": "1.89.2"
   }
 }
